Rename PageTemplatesLegacy to PageTemplates
The default template list component was misleadingly named "legacy"
when it's the one actually used. Rename it to `index.js` (default)
and move the experimental template activation variant to
`index-activation.js`.
